Substring
Substring in java is the subset of  main string specified by start and end indices .  
String started from index 0 .   
s1 reference variable stored string and print as it is.  
In above program when we print s3 string will print all the character after index 4. 
And at the time of printing s4 all the character print between index 3 and  11.